+++ src/LiberisLabs.DogStatsD.Interceptors.Tests/DogStatsdInterceptorTests/DogStatsdInterceptorTestsForException.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -6,32 +6,42 @@ using Castle.DynamicProxy;
 using LiberisLabs.DogStatsD.Interceptors.TaskInterceptors;
 using Moq;
 using NUnit.Framework;
-using IInterceptor = LiberisLabs.DogStatsD.Interceptors.Interceptors.IInterceptor;
 
-namespace LiberisLabs.DogStatsD.Interceptors.Tests.DogStatdInterceptorTests
+namespace LiberisLabs.DogStatsD.Interceptors.Tests.DogStatsdInterceptorTests
 {
     [TestFixture]
-    public class DogStatdInterceptorTests
+    public class DogStatsdInterceptorTestsForException
     {
         private Mock<ITaskInterceptor> _interceptor1;
         private Mock<ITaskInterceptor> _interceptor2;
+        private Exception _exception;
 
         [SetUp]
-        public void GivenADogStatdInterceptor_WhenInterceptASuccessfulCall()
+        public void GivenADogStatsdInterceptor_WhenInterceptASuccessfulCall()
         {
+            _exception = new Exception();
             _interceptor1 = new Mock<ITaskInterceptor>();
             _interceptor2 = new Mock<ITaskInterceptor>();
 
             var factory = new Mock<IInterceptorFactory>();
             factory.Setup(x => x.CreateInterceptors(It.IsAny<MethodInfo>(), It.IsAny<MethodInfo>()))
-                .Returns(new List<ITaskInterceptor>() {_interceptor1.Object, _interceptor2.Object});
+                .Returns(new List<ITaskInterceptor>() { _interceptor1.Object, _interceptor2.Object });
 
-            var interceptor = new DogStatdInterceptor(factory.Object);
+            var interceptor = new DogStatsdInterceptor(factory.Object);
 
             var invocation = new Mock<IInvocation>();
-            invocation.Setup(x => x.ReturnValue).Returns(1);
+            invocation.Setup(x => x.Proceed())
+                .Throws(_exception);
 
-            interceptor.Intercept(invocation.Object);
+            try
+            {
+                interceptor.Intercept(invocation.Object);
+
+            }
+            catch (Exception)
+            {
+                // ignored
+            }
         }
 
         [Test]
@@ -42,17 +52,17 @@ namespace LiberisLabs.DogStatsD.Interceptors.Tests.DogStatdInterceptorTests
         }
 
         [Test]
-        public void ThenAllInterceptorsOnExitIsCalledOnce()
+        public void ThenInterceptorsOnExitIsNeverCalledOnce()
         {
-            _interceptor1.Verify(x => x.OnExit(), Times.Once);
-            _interceptor2.Verify(x => x.OnExit(), Times.Once);
+            _interceptor1.Verify(x => x.OnExit(), Times.Never);
+            _interceptor2.Verify(x => x.OnExit(), Times.Never);
         }
 
         [Test]
-        public void ThenInterceptorsOnExceptionIsNeverCalled()
+        public void ThenAllInterceptorsOnExceptionIsCalled()
         {
-            _interceptor1.Verify(x => x.OnException(It.IsAny<Exception>()), Times.Never);
-            _interceptor2.Verify(x => x.OnException(It.IsAny<Exception>()), Times.Never);
+            _interceptor1.Verify(x => x.OnException(_exception), Times.Once);
+            _interceptor2.Verify(x => x.OnException(_exception), Times.Once);
         }
 
         [Test]
