+++ src/LiberisLabs.DogStatsD.Interceptors.Tests/LiberisLabs.DogStatsD.Interceptors.Tests.csproj
@@ -52,9 +52,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DogStatdInterceptorTests\DogStatdInterceptorTests.cs" />
-    <Compile Include="DogStatdInterceptorTests\DogStatdInterceptorTestsForException.cs" />
-    <Compile Include="DogStatdInterceptorTests\DogStatdInterceptorTestsForMethodsReturningTasks.cs" />
+    <Compile Include="DogStatsdInterceptorTests\DogStatsdInterceptorTests.cs" />
+    <Compile Include="DogStatsdInterceptorTests\DogStatsdInterceptorTestsForException.cs" />
+    <Compile Include="DogStatsdInterceptorTests\DogStatsdInterceptorTestsForMethodsReturningTasks.cs" />
     <Compile Include="Helpers\MethodInfoBuilder.cs" />
     <Compile Include="InterceptorFactoryTests.cs" />
     <Compile Include="Interceptors\MonitorInterceptorAdapterTests\MonitorInterceptorAdapterTestsForCanIntercept.cs" />
