+++ src/LiberisLabs.DogStatsD.Interceptors/LiberisLabs.DogStatsD.Interceptors.csproj
@@ -61,7 +61,7 @@
     <Compile Include="TaskInterceptors\TaskTimerInterceptor.cs" />
     <Compile Include="Interceptors\TimerInterceptor.cs" />
     <Compile Include="Monitors\InstrumentMonitor.cs" />
-    <Compile Include="DogStatdInterceptor.cs" />
+    <Compile Include="DogStatsdInterceptor.cs" />
     <Compile Include="DogStatsdWrapper.cs" />
     <Compile Include="Interceptors\IInterceptor.cs" />
     <Compile Include="IDogStatsd.cs" />
